+++ frontend/src/pages/QueryDocuments/components/QuerySelectVDB.tsx
@@ -36,6 +36,7 @@ export default function QuerySelectVDB() {
   // update gvdbs_retr_params from group_vdbs row if necessary
   useEffect(() => {
     if (docTaskOptionsStore.needReload) return;
+    if (!current.doc_task_id) return;
     if (!current.gvdbs_id || current.gvdbs_id === -1) return;
     const gvdbs_row = docTaskOptionsStore.gvdbs_id__row[current.gvdbs_id];
     const gvdbs_retr_params_str = gvdbs_row?.gvdbs_retr_params;
@@ -47,11 +48,17 @@ export default function QuerySelectVDB() {
     if (gvdbs_retr_filters_str) {
       defGVDBsRetrFiltersStore.setDataFromString(gvdbs_retr_filters_str);
       curGVDBsRetrFiltersStore.reset();
+      curGVDBsRetrFiltersStore.loadFromRFHistory(
+        current.doc_task_id,
+        current.gvdbs_id,
+        JSON.parse(gvdbs_retr_filters_str),
+      );
     }
   }, [
     docTaskOptionsStore.needReload,
     defGVDBsRetrParamsStore.isLoaded,
     defGVDBsRetrFiltersStore.isLoaded,
+    current.doc_task_id,
     current.gvdbs_id,
   ]);